Se obtiene la nota del usuario
